squash migrations into one commit
- Combines all previous migrations into one base migration
- Removes old migration files
- Maintains same database structure
- No functional changes
