+++ alembic/alembic.ini
@@ -3,7 +3,7 @@ script_location = .
 file_template = %%(year)d%%(month).2d%%(day).2d_%%(hour).2d%%(minute).2d-%%(slug)s
 
 prepend_sys_path = ..
-version_path_separator = os  # Use os.pathsep. Default configuration used for new projects.
+version_path_separator = os
 
 # URL
 sqlalchemy.url = ${env:CHAI_DATABASE_URL}
