--- alembic/versions/20241018_0815-check_default_server_types.py
@@ -1,614 +1,0 @@
-"""check default server types
-
-Revision ID: 2481138a729a
-Revises: d183dcc4bdc8
-Create Date: 2024-10-18 08:15:28.100836
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "2481138a729a"
-down_revision: Union[str, None] = "d183dcc4bdc8"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    op.alter_column(
-        "dependencies",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "dependencies",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "dependencies",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "depends_on_types",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "depends_on_types",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "depends_on_types",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "licenses",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "licenses",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "licenses",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "load_history",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "load_history",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "load_history",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_managers",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_managers",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_managers",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_urls",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_urls",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_urls",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "packages",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "packages",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "packages",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "sources",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "sources",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "sources",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "url_types",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "url_types",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "url_types",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "urls",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "urls",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "urls",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_packages",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_packages",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_packages",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_versions",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_versions",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_versions",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "users",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "users",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "users",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "versions",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=sa.text("uuid_generate_v4()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "versions",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "versions",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=sa.text("now()"),
-        existing_nullable=False,
-    )
-
-
-def downgrade() -> None:
-    op.alter_column(
-        "versions",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "versions",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "versions",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "users",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "users",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "users",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_versions",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_versions",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_versions",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_packages",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_packages",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "user_packages",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "urls",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "urls",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "urls",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "url_types",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "url_types",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "url_types",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "sources",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "sources",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "sources",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "packages",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "packages",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "packages",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_urls",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_urls",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_urls",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_managers",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_managers",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "package_managers",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "load_history",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "load_history",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "load_history",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "licenses",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "licenses",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "licenses",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "depends_on_types",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "depends_on_types",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "depends_on_types",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "dependencies",
-        "updated_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "dependencies",
-        "created_at",
-        existing_type=postgresql.TIMESTAMP(),
-        server_default=None,
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "dependencies",
-        "id",
-        existing_type=sa.UUID(),
-        server_default=None,
-        existing_nullable=False,
-    )
